bpf: Allow to use 24 bits for security identities
This changes the BPF datapath to use 32 bit integers to handle security
identities. It also adjusts the skb mark handling to split up the identity and
use part of the existing magic marker to store the new extended 8 bits of the
identity.

This commit does not extend the the allocator to use additional identity space
yet.

+++ daemon/endpoint.go
@@ -868,7 +868,7 @@ func (d *Daemon) OnIPIdentityCacheChange(modType ipcache.CacheModification, oldI
 	switch modType {
 	case ipcache.Upsert:
 		value := ipCacheBPF.RemoteEndpointInfo{
-			SecurityIdentity: uint16(newIPIDPair.ID),
+			SecurityIdentity: uint32(newIPIDPair.ID),
 		}
 
 		if newIPIDPair.HostIP != nil {
